+++ lucene/analysis/common/src/test/org/apache/lucene/analysis/util/TestCSVUtil.java
@@ -14,10 +14,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.lucene.analysis.ja;
+package org.apache.lucene.analysis.util;
 
 import java.io.IOException;
-import org.apache.lucene.analysis.ja.dict.CSVUtil;
 import org.apache.lucene.tests.util.LuceneTestCase;
 
 /*
