+++ lucene/analysis/kuromoji/src/test/org/apache/lucene/analysis/ja/dict/TestUnknownDictionary.java
@@ -16,6 +16,7 @@
  */
 package org.apache.lucene.analysis.ja.dict;
 
+import org.apache.lucene.analysis.util.CSVUtil;
 import org.apache.lucene.tests.util.LuceneTestCase;
 import org.junit.Test;
